requirements-test.txt
# Test dependencies for Claude Development Framework
# Install with: pip install -r requirements-test.txt

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0  # Parallel test execution

# YAML parsing (for agent metadata)
pyyaml>=6.0

# Assertion helpers
pytest-assert-utils>=0.3.1

# Test reporting
pytest-html>=4.1.1

# Code coverage
coverage[toml]>=7.4.0

# Performance testing
pytest-benchmark>=4.0.0

# Optional: For mocking HTTP requests in integration tests
responses>=0.24.1
requests-mock>=1.11.0
